requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
passlib==1.7.2
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.1
